Fix error and weirdness related to `gardenspreadRate` config
The value of this config option does not correspond to the odds of a garden spawning each tick in a way that is easy to understand. Most allowed values result in very low spread rates, with spread rates only noticeably increasing with values above 60 (and they increase very rapidly from there on).

The comment near this config option reads,

`# Garden spread rate. 100 means a garden spawns every tick. 1 means a garden spawns with a probability of 1% per tick.`

However, it is at 0 that gardens will spawn with a probability of 1% (or 1:100). It is at 99 that a garden will spawn each tick. Setting the config option to 100 causes an `IllegalArgumentException`.

Reading the current code for the garden spawning:

`if (config.enablegardenSpread && rand.nextInt(100 - config.gardenspreadRate) == 0)`

With the config at 1, the odds of a garden spawning each tick is 1:99. With the config at 30 (the default), the odds are 1:70, which is not much greater than 1%. To get the odds of spawning up to 1:20, the config option must be set to 80; to get odds of 1:2, the config option must be set to 98.

This change will make the config option actually result in an X% probability of spawning a garden and fix the aforementioned exception.